# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Domicilia Core API
  version: 0.1.0
  description: |
    Contrato de la API de domicilia-core. Es la FUENTE DE VERDAD para cualquier cliente (el frontend, el agente de IA):
    los clientes tipados se generan de este archivo.

    Convenciones:
    - Identificadores: UUID como texto.
    - Errores: `application/problem+json` (RFC 9457) con `detail` y `request_id`.
    - Listados que crecen (conversaciones, mensajes, contactos): paginación por CURSOR (`limit` + `cursor`); la respuesta
      trae `next_cursor` (null en la última página). El cursor es opaco.
    - Fechas: RFC 3339 en UTC.
    - Cambios parciales (PATCH): un campo ausente no se toca; `null` deja el campo en blanco.

    ESTADO: este contrato cubre por ahora el módulo de WhatsApp (bandejas, contactos, conversaciones y mensajes) y su
    webhook, el módulo de catálogo (categorías, productos, variantes y modificadores), el de pedidos (carrito y su
    máquina de estados), el de pagos (cobro por ePayco y su webhook) y el de promociones (cupones de descuento). El
    resto de la API (usuarios, organizaciones, planes, roles) se documenta aparte; una prueba
    verifica que lo documentado aquí es exactamente lo que el servidor sirve, y valida las respuestas reales contra
    estos esquemas.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /v1/organizations/{org_id}/inboxes:
    get:
      tags: [Bandejas]
      operationId: listInboxes
      summary: Lista las bandejas de la organización
      description: Requiere `org.inbox.read`. Las archivadas solo las ve quien tiene `org.inbox.manage` y las pide con `include_archived=true`.
      parameters:
        - $ref: '#/components/parameters/OrgID'
        - name: include_archived
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Bandejas
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Inbox' } }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/inboxes/whatsapp:
    post:
      tags: [Bandejas]
      operationId: connectWhatsAppInbox
      summary: Conecta un número de WhatsApp
      description: |
        Requiere `org.inbox.manage`. Valida las credenciales CONTRA Meta antes de guardar nada: el número debe pertenecer a la
        cuenta y el token debe poder verlo. La primera bandeja es del plan base; más de una exige la función `multi_inbox` (402).
        Responde 503 si el servidor no tiene configurado el cifrado de secretos.
      parameters:
        - $ref: '#/components/parameters/OrgID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConnectWhatsAppRequest' }
      responses:
        '201':
          description: Bandeja conectada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Inbox' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/inboxes/{inbox_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/InboxID'
    get:
      tags: [Bandejas]
      operationId: getInbox
      summary: Lee una bandeja
      responses:
        '200':
          description: Bandeja
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Inbox' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Bandejas]
      operationId: renameInbox
      summary: Cambia el nombre de la bandeja
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenameInboxRequest' }
      responses:
        '200':
          description: Bandeja actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Inbox' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Bandejas]
      operationId: archiveInbox
      summary: Desconecta la bandeja
      description: Archiva la bandeja y libera el número (se puede reconectar). Las conversaciones se conservan.
      responses:
        '204': { description: Desconectada }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/inboxes/{inbox_id}/credentials:
    put:
      tags: [Bandejas]
      operationId: rotateInboxToken
      summary: Cambia el token de Meta de la bandeja
      description: Se valida contra Meta con el mismo número antes de reemplazar el guardado. Reactiva un canal en `needs_reauth`.
      parameters:
        - $ref: '#/components/parameters/OrgID'
        - $ref: '#/components/parameters/InboxID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RotateCredentialsRequest' }
      responses:
        '200':
          description: Bandeja con el canal reactivado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Inbox' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/inboxes/{inbox_id}/refresh:
    post:
      tags: [Bandejas]
      operationId: refreshInbox
      summary: Vuelve a leer de Meta los datos del número
      description: Actualiza calidad y nivel de mensajería. Si Meta rechaza el token, el canal pasa a `needs_reauth`.
      parameters:
        - $ref: '#/components/parameters/OrgID'
        - $ref: '#/components/parameters/InboxID'
      responses:
        '200':
          description: Bandeja actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Inbox' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/contacts:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    get:
      tags: [Contactos]
      operationId: listContacts
      summary: Lista contactos (los más nuevos primero)
      parameters:
        - name: q
          in: query
          description: Busca en nombre, teléfono y correo. Los comodines `%` y `_` se toman literales.
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Página de contactos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContactPage' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Contactos]
      operationId: createContact
      summary: Crea un contacto
      description: El teléfono debe traer el código de país y es único por organización.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateContactRequest' }
      responses:
        '201':
          description: Contacto creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/contacts/{contact_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/ContactID'
    get:
      tags: [Contactos]
      operationId: getContact
      summary: Lee un contacto
      responses:
        '200':
          description: Contacto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Contactos]
      operationId: updateContact
      summary: Cambia un contacto (parcial)
      description: 'El teléfono no se edita: es la identidad del contacto.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateContactRequest' }
      responses:
        '200':
          description: Contacto actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de auth-domicilia (GoTrue).
  parameters:
    OrgID:
      name: org_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    InboxID:
      name: inbox_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    ContactID:
      name: contact_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      description: Filas por página (1 a 100, por omisión 30).
      schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
    Cursor:
      name: cursor
      in: query
      description: Valor de `next_cursor` de la página anterior.
      schema: { type: string }
  responses:
    Problem:
      description: |
        Error. Códigos habituales: 401 sin credenciales; 402 el plan no incluye la función; 403 sin permiso o organización
        suspendida; 404 no existe (o archivada); 409 choca con el estado actual; 422 entrada inválida; 502 falló Meta;
        503 función sin configurar.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
...
    WhatsAppChannel:
      type: object
      description: El número conectado. No trae secretos.
      required: [phone_number_id, display_phone, verified_name, quality_rating, messaging_tier, status, waba_id, token_status, last_checked_at]
      properties:
        phone_number_id: { type: string, description: Id del número en Meta. }
        display_phone: { type: string }
        verified_name: { type: string }
        quality_rating: { type: string, description: 'GREEN, YELLOW o RED (de Meta).' }
        messaging_tier: { type: string, description: 'Nivel de mensajería de Meta (p. ej. TIER_1K).' }
        status: { type: string, enum: [connected, needs_reauth] }
        waba_id: { type: string }
        token_status: { type: string, enum: [valid, invalid] }
        last_checked_at: { type: string, format: date-time, nullable: true }
    Inbox:
      type: object
      required: [id, organization_id, name, channel_type, archived_at, created_at, updated_at, whatsapp]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        name: { type: string }
        channel_type: { type: string, enum: [whatsapp] }
        archived_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        whatsapp:
          nullable: true
          description: Null si la bandeja está archivada (el número se liberó).
          allOf:
            - $ref: '#/components/schemas/WhatsAppChannel'
    ConnectWhatsAppRequest:
      type: object
      required: [name, waba_id, phone_number_id, access_token]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        waba_id: { type: string, pattern: '^[0-9]{5,32}$' }
        phone_number_id: { type: string, pattern: '^[0-9]{5,32}$' }
        access_token: { type: string, maxLength: 1024, description: 'Se guarda cifrado; nunca se devuelve.' }
        app_secret:
          type: string
          description: Solo si el negocio usa su PROPIA app de Meta. Con la app de la plataforma va vacío.
    RenameInboxRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
    RotateCredentialsRequest:
      type: object
      required: [access_token]
      properties:
        access_token: { type: string, maxLength: 1024 }
    Contact:
      type: object
      required: [id, organization_id, phone, name, email, custom_attributes, blocked, source, last_activity_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        phone: { type: string, description: 'E.164, con "+".', example: '+573001234567' }
        name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        custom_attributes:
          type: object
          description: 'Hasta 50 atributos de texto (500 caracteres), número, booleano o null.'
          additionalProperties: true
        blocked: { type: boolean }
        source: { type: string, enum: [inbound, manual, import, order, api] }
        last_activity_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateContactRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, description: 'Con código de país; se aceptan espacios, guiones y paréntesis.' }
        name: { type: string, maxLength: 255, nullable: true }
        email: { type: string, nullable: true }
        custom_attributes: { type: object, additionalProperties: true }
    UpdateContactRequest:
      type: object
      description: Un campo ausente no se toca; en `name` y `email`, una cadena vacía o null lo deja en blanco.
      properties:
        name: { type: string, maxLength: 255, nullable: true }
        email: { type: string, nullable: true }
        custom_attributes: { type: object, additionalProperties: true }
        blocked: { type: boolean }
    ContactPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Contact' } }
        next_cursor: { type: string, nullable: true }
...